% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{nomnoml-syntax}
\alias{nomnoml-syntax}
\alias{`nomnoml-syntax`}
\title{nomnoml_syntax}
\description{
Summary of nomnoml syntax.
}
\section{Association Types}{


\verb{association -}

\verb{association ->}

\verb{association <->}

\verb{dependency -->}

\verb{dependency <-->}

\verb{generalization -:>}

\verb{generalization <:-}

\verb{implementation --:>}

\verb{implementation <:--}

\verb{composition +-}

\verb{composition +->}

\verb{aggregation o-}

\verb{aggregation o->}

\verb{note --}

\verb{hidden -/-}
}

\section{Classifier Types}{


\verb{[name]}

\verb{[<abstract> name]}

\verb{[<instance> name]}

\verb{[<note> name]}

\verb{[<reference> name]}

\verb{[<package> name]}

\verb{[<frame> name]}

\verb{[<database> name]}

\verb{[<start> name]}

\verb{[<end> name]}

\verb{[<state> name]}

\verb{[<choice> name]}

\verb{[<input> name]}

\verb{[<sender> name]}

\verb{[<receiver> name]}

\verb{[<transceiver> name]}

\verb{[<actor> name]}

\verb{[<usecase> name]}

\verb{[<label> name]}

\verb{[<hidden> name]}
}

\section{Directives}{


\verb{#arrowSize: 1}

\verb{#bendSize: 0.3}

\verb{#direction: down | right}

\verb{#gutter: 5}

\verb{#edgeMargin: 0}

\verb{#edges: hard | rounded}

\verb{#fill: #eee8d5; #fdf6e3}

\verb{#fillArrows: false}

\verb{#font: Calibri}

\verb{#fontSize: 12}

\verb{#leading: 1.25}

\verb{#lineWidth: 3}

\verb{#padding: 8}

\verb{#spacing: 40}

\verb{#stroke: #33322E}

\verb{#title: filename}

\verb{#zoom: 1}
}

\section{Custom Classifier Styles}{


A directive that starts with \code{.} define a classifier style.
The style is written as a space separated list of modifiers and
key/value pairs.

\verb{#.box: fill=#8f8 dashed}

\verb{#.blob: visual=ellipse}

\verb{[<box> GreenBox]}

\verb{[<blob> HideousBlob]}
\subsection{Available key/value pairs are:}{

\verb{fill=(any css color)}

\verb{stroke=(any css color)}

\code{align=center}

\code{align=left}

\code{direction=right}

\code{direction=down}

\code{visual=actor}

\code{visual=class}

\code{visual=database}

\code{visual=ellipse}

\code{visual=end}

\code{visual=frame}

\code{visual=hidden}

\code{visual=input}

\code{visual=none}

\code{visual=note}

\code{visual=package}

\code{visual=receiver}

\code{visual=rhomb}

\code{visual=roundrect}

\code{visual=sender}

\code{visual=start}

\code{visual=transceiver}
}

\subsection{Available modifiers are:}{

\code{bold}

\code{underline}

\code{italic}

\code{dashed}

\code{empty}
}
}

